requirements.txt
filelock==3.25.0
fsspec==2026.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
# packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049099114/work
pillow>=10.0
PyYAML>=6.0
numpy>=1.26
setuptools==80.10.2
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
wheel==0.46.3
matplotlib
tqdm
pandas
